+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.0
+10 june 2017
+
+* New services: geo, http, pkg, key, router, scroll, user
+* New parent service: subscribe
+
 ## v0.0.1
 8 june 2017